+++ StrandiOS/LiveActivity/NOOPLiveActivityBundle.swift
@@ -1,0 +1,10 @@
+import WidgetKit
+import SwiftUI
+import ActivityKit
+
+@main
+struct NOOPLiveActivityBundle: WidgetBundle {
+    var body: some Widget {
+        WorkoutLiveActivity()
+    }
+}